+++ rl2/agents/architectures/common/normalization.py
@@ -1,0 +1,31 @@
+"""
+Implements normalization layers for RL^2 agents.
+"""
+
+import torch as tc
+
+
+class LayerNorm(tc.nn.Module):
+    """
+    Layer Normalization.
+    """
+    def __init__(self, units):
+        super().__init__()
+        self._units = units
+        self._g = tc.nn.Parameter(tc.ones(units, device='cpu'))
+        self._b = tc.nn.Parameter(tc.zeros(units, device='cpu'))
+
+    def forward(self, inputs, eps=1e-8):
+        mu = tc.mean(inputs, dim=-1, keepdim=True)
+        centered = inputs - mu
+        sigma2 = tc.mean(tc.square(centered), dim=-1, keepdim=True)
+        sigma = tc.sqrt(sigma2 + eps)
+        standardized = centered / sigma
+
+        g, b = self._g, self._b
+        while len(list(g.shape)) < len(list(inputs.shape)):
+            g = g.unsqueeze(0)
+            b = b.unsqueeze(0)
+
+        scaled = g * standardized + b
+        return scaled
